pyproject.toml
[tool.poetry]
name = "blog-manager-action"
version = "0.1.0"
description = ""
authors = ["WoolDoughnut310 <wooldoughnutspi@outlook.com>"]
readme = "README.md"
packages = [{include = "blog_manager_action"}]

[tool.poetry.dependencies]
python = "^3.9"
pygithub = "^1.59.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-frontmatter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
